更新到1.19.4版本的第一个版本

+++ src/main/java/com/blue0666/carpetblueaddition/event/soundlistenersystem/SoundEventManager.java
@@ -28,7 +28,7 @@ public class SoundEventManager {
     }
 
     public static void handleEntitySound(Entity entity, int soundLevel) {
-        BlockPos interactPos = new BlockPos(entity.getX() + 0.5, entity.getY(), entity.getZ() + 0.5);
+        BlockPos interactPos = new BlockPos((int)(entity.getX() + 0.5), (int)entity.getY(), (int)(entity.getZ() + 0.5));
         World world = entity.getEntityWorld();
         Box searchArea = createBoxArea(interactPos, getListeningRadius());
